pyproject.toml
[tool.poetry]
name = "u-t2"
version = "0.1.0"
description = ""
authors = ["HDAlbarenque <darioalbarenque@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
mega-py = "^1.0.8"
openpyxl = "^3.1.2"
firebase-admin = "^6.4.0"
icecream = "^2.1.3"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
